Add nixos subcommand to apply
